requirements.txt
Flask
transformers
torch
torch_directml
google-search-results
openai
python-dotenv
